+++ public/components/health-check/services/check-index-pattern/check-index-pattern-object.service.ts
@@ -34,7 +34,7 @@ export const checkIndexPatternObjectService =  async (appConfig, checkLogger: Ch
   checkLogger.info(`Found default index pattern with title [${defaultPatternId}]: ${indexPatternDefaultFound ? 'yes' : 'no'}`);
 
   if (!indexPatternDefaultFound && defaultPatternId) {
-    // if no valid index patterns are found we try to create the wazuh-alerts-*
+    // if no valid index patterns are found we try to create the ams-alerts-*
     try {
       checkLogger.info(`Checking if index pattern [${defaultPatternId}] exists...`);
       const existDefaultIndexPattern = await SavedObject.getExistingIndexPattern(defaultPatternId);