+++ server/controllers/wazuh-elastic.ts
@@ -340,7 +340,7 @@ export class WazuhElasticCtrl {
   /**
    * Replaces visualizations main fields to fit a certain pattern.
    * @param {Array<Object>} app_objects Object containing raw visualizations.
-   * @param {String} id Index-pattern id to use in the visualizations. Eg: 'wazuh-alerts'
+   * @param {String} id Index-pattern id to use in the visualizations. Eg: 'ams-alerts'
    */
   async buildVisualizationsRaw(app_objects, id, namespace = false) {
     try {
@@ -392,7 +392,7 @@ export class WazuhElasticCtrl {
             );
           } else {
             aux_source.kibanaSavedObjectMeta.searchSourceJSON = defaultStr.replace(
-              /wazuh-alerts/g,
+              /ams-alerts/g,
               id
             );
           }
@@ -401,7 +401,7 @@ export class WazuhElasticCtrl {
         // Replace index-pattern for selector visualizations
         if (typeof (aux_source || {}).visState === 'string') {
           aux_source.visState = aux_source.visState.replace(
-            /wazuh-alerts/g,
+            /ams-alerts/g,
             id
           );
         }
@@ -427,7 +427,7 @@ export class WazuhElasticCtrl {
   /**
    * Replaces cluster visualizations main fields.
    * @param {Array<Object>} app_objects Object containing raw visualizations.
-   * @param {String} id Index-pattern id to use in the visualizations. Eg: 'wazuh-alerts'
+   * @param {String} id Index-pattern id to use in the visualizations. Eg: 'ams-alerts'
    * @param {Array<String>} nodes Array of node names. Eg: ['node01', 'node02']
    * @param {String} name Cluster name. Eg: 'wazuh'
    * @param {String} master_node Master node name. Eg: 'node01'
@@ -447,7 +447,7 @@ export class WazuhElasticCtrl {
       for (const element of app_objects) {
         // Stringify and replace index-pattern for visualizations
         aux_source = JSON.stringify(element._source);
-        aux_source = aux_source.replace(/wazuh-alerts/g, id);
+        aux_source = aux_source.replace(/ams-alerts/g, id);
         aux_source = JSON.parse(aux_source);
 
         // Bulk source