+++ server/integration-files/visualizations/agents/agents-audit.ts
@@ -55,11 +55,11 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON: JSON.stringify({
-          index: 'wazuh-alerts',
+          index: 'ams-alerts',
           filter: [
             {
               meta: {
-                index: 'wazuh-alerts',
+                index: 'ams-alerts',
                 negate: false,
                 disabled: false,
                 alias: null,
@@ -135,11 +135,11 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON: JSON.stringify({
-          index: 'wazuh-alerts',
+          index: 'ams-alerts',
           filter: [
             {
               meta: {
-                index: 'wazuh-alerts',
+                index: 'ams-alerts',
                 negate: false,
                 disabled: false,
                 alias: null,
@@ -215,11 +215,11 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON: JSON.stringify({
-          index: 'wazuh-alerts',
+          index: 'ams-alerts',
           filter: [
             {
               meta: {
-                index: 'wazuh-alerts',
+                index: 'ams-alerts',
                 type: 'phrases',
                 key: 'rule.id',
                 value: '80781, 80787',
@@ -301,11 +301,11 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON: JSON.stringify({
-          index: 'wazuh-alerts',
+          index: 'ams-alerts',
           filter: [
             {
               meta: {
-                index: 'wazuh-alerts',
+                index: 'ams-alerts',
                 negate: false,
                 disabled: false,
                 alias: null,
@@ -366,7 +366,7 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON: JSON.stringify({
-          index: 'wazuh-alerts',
+          index: 'ams-alerts',
           filter: [],
           query: { query: '', language: 'lucene' },
         }),
@@ -404,7 +404,7 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON: JSON.stringify({
-          index: 'wazuh-alerts',
+          index: 'ams-alerts',
           filter: [],
           query: { query: '', language: 'lucene' },
         }),
@@ -507,7 +507,7 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON: JSON.stringify({
-          index: 'wazuh-alerts',
+          index: 'ams-alerts',
           filter: [],
           query: { query: '', language: 'lucene' },
         }),
@@ -545,7 +545,7 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON:
-          '{"index":"wazuh-alerts","filter":[],"query":{"query":"","language":"lucene"}}',
+          '{"index":"ams-alerts","filter":[],"query":{"query":"","language":"lucene"}}',
       },
     },
     _type: 'visualization',
@@ -629,7 +629,7 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON: JSON.stringify({
-          index: 'wazuh-alerts',
+          index: 'ams-alerts',
           filter: [],
           query: { query: '', language: 'lucene' },
         }),