+++ server/integration-files/visualizations/agents/agents-pm.ts
@@ -108,7 +108,7 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON: JSON.stringify({
-          index: 'wazuh-alerts',
+          index: 'ams-alerts',
           filter: [],
           query: { query: '', language: 'lucene' },
         }),
@@ -162,7 +162,7 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON: JSON.stringify({
-          index: 'wazuh-alerts',
+          index: 'ams-alerts',
           query: { query: '', language: 'lucene' },
           filter: [],
         }),
@@ -250,7 +250,7 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON: JSON.stringify({
-          index: 'wazuh-alerts',
+          index: 'ams-alerts',
           filter: [],
           query: { query: '', language: 'lucene' },
         }),
@@ -320,7 +320,7 @@ export default [
       version: 1,
       kibanaSavedObjectMeta: {
         searchSourceJSON: JSON.stringify({
-          index: 'wazuh-alerts',
+          index: 'ams-alerts',
           filter: [],
           query: { query: '', language: 'lucene' },
         }),